requirements.txt
tqdm==4.47.0
nltk==3.5
stanza==1.0.1
numpy==1.19.0
pandas==1.1.0
gensim==3.8.3
pyemd==0.5.1
scikit-learn==0.23.2
lightgbm==2.3.1
PyStemmer==2.0.1
requests==2.24.0
